fixed backend to pass tests

+++ server/controllers/controller.js
@@ -30,7 +30,9 @@ const createNote = async (req, res) => {
 const getNote = async (req, res) => {
   try {
     const note = await Note.findById(req.params.id);
-    res.status(200).json(note);
+    if(note) res.status(200).json(note);
+    else res.status(404).json({ error: 'No note found with provided ID' })
+    
     return note;
   } catch (error) {
     res.status(500).json({ error });
@@ -44,6 +46,10 @@ const updateNote = async (req, res) => {
 
   try {
     const note = await Note.findById(req.params.id);
+    if (!note) {
+      res.status(404).json({ error: 'No note found with provided ID' })
+      return 
+    }
     note.title = req.body.title;
     note.body = req.body.body;
     note.date = req.body.date;
@@ -51,9 +57,13 @@ const updateNote = async (req, res) => {
     note.user = req.body.user;
     note.favorite = req.body.favorite;
     console.log('note in the controller after change', note);
-    await note.save();
-    res.status(200).json(note);
-    return note;
+    try {
+      await note.save();
+      res.status(200).json(note);
+    } catch (e) {
+      console.error(e);
+      res.status(409).json({ error: 'Failed to update note' })
+    }
   } catch (error) {
     res.status(500).json({ error });
     return error;
@@ -61,14 +71,14 @@ const updateNote = async (req, res) => {
 };
 
 const deleteNote = async (req, res) => {
-  console.log('deleteNote called in the controller', req.params.id);
   try {
     const note = await Note.findById(req.params.id);
-    console.log(note);
+    if (!note) { 
+      res.status(404).json({ error: 'Could not find note' })
+      return
+    }
     await Note.deleteOne(note);
     res.status(200).json(note);
-    console.log('sucess');
-    return note;
   } catch (error) {
     res.status(500).json({ error });
     return error;
